Move strategies to `strategies.py`
